+++ example/lib/thermal_printer_sample_screen.dart
@@ -1,7 +1,6 @@
 import 'dart:developer';
 import 'package:blue_thermal_helper/blue_thermal_helper.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 /// Thermal Printer Sample Screen
 ///